Remove unnecessary CSS file and update config for production env
Deleted the 'page.module.css' as it was no longer being used, reducing the codebase size. Changes were also made to 'next.config.js', 'package.json' and '.gitignore' to accommodate for production environment. These changes include setting a base path for deployment in 'next.config.js', adding a command to create a '.nojekyll' file in 'package.json' to avoid potential build problems with Jekyll and Github Pages, and updating '.gitignore' to ignore the '.idea' directory for JetBrains IDEs like IntelliJ and WebStorm. Also, the 'README.md' is simplified to reflect changes and purpose of the project.